rnd/document_classification/requirements.txt
torch==2.0.0
torchvision==0.15.0
transformers==4.30.0
Pillow==9.0.0
scikit-learn==1.0.0
accelerate==0.20.0
datasets==2.10.0
evaluate==0.4.0